+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Make interchangeable with `<Image>` component and intrinsic `<img>` tag ([#21](https://github.com/speee/jsx-slack/pull/21))
 
+### Removed
+
+- Remove deprecated `<Block>` component ([#22](https://github.com/speee/jsx-slack/pull/22))
+
 ## v0.4.3 - 2019-05-15
 
 ### Fixed